# trvCollection.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "trvCollection",
    "description": "# Introduction\nSimple express API with 5 main functionalities.\n * Managing rowing days appointments.\n * User authentification.\n * Managing news of the website.\n * Managing user data.\n * Enabling pdf download.\n",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/new_user": {
      "post": {
        "description": "Save a new user",
        "summary": "http://localhost:3000/api/new_user",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/new_user",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1new_user_request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1new_user"
            },
            "examples": {
              "application/json": {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTAsImlhdCI6MTYwNzQ1OTY2NSwiZXhwIjoxNjA3NDY2ODY1fQ.vhVXS5t06x4B1d2lxmGM6MEn0A7EFl722WYA3rsH3XU",
                "user": {
                  "id": 10,
                  "name": "admin",
                  "email": "adminaaaa@mail.com",
                  "role": "ADMIN"
                }
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "227"
              },
              "etag": {
                "type": "string",
                "default": "W/\"e3-7jB0/brD7gPDNmU0i26ot7iIN8U\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 20:34:25 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "User already exists!"
              }
            }
          }
        }
      }
    },
    "/new_user_reg": {
      "post": {
        "description": "Register a new user",
        "summary": "http://localhost:3000/api/new_user_reg",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/new_user_reg",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1new_user_request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Get all users ",
        "summary": "http://localhost:3000/api/users",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/users",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1user"
              }
            },
            "examples": {
              "application/json": [
                {
                  "id": 11,
                  "name": "admin",
                  "email": "adwwwminaaaa@mail.com",
                  "password": "$2a$10$kmlURv.RNIYegcJEdK0C6.MtCJ09FIDS.iGMrhwdNbFJkbKlVcmha",
                  "role": "ADMIN",
                  "created_at": null
                },
                {
                  "id": 10,
                  "name": "admin",
                  "email": "adminaaaa@mail.com",
                  "password": "$2a$10$I8wS/0x4/kKx24Y4/lNx3.IIfYt41R7Fo/tunI8BICHuPxPT2DwUG",
                  "role": "ADMIN",
                  "created_at": null
                },
                {
                  "id": 4,
                  "name": "test editor",
                  "email": "testeditor@mail.com",
                  "password": "$2a$10$H6EmQWcUtMQMEMQCh6BQR.bkKJTTPhprXiSrxdPujG7PCrP7V/HqK",
                  "role": "EDITOR",
                  "created_at": null
                },
                {
                  "id": 9,
                  "name": "Tony Vercetti",
                  "email": "tony@mail.com",
                  "password": "$2a$10$8hi7SETeQJwdSfPjHUIS5.L/pzE0OoGkvJE6XguagJ3ysQUR0IFnq",
                  "role": "ADMIN",
                  "created_at": null
                },
                {
                  "id": 8,
                  "name": "admin",
                  "email": "admin@mail.com",
                  "password": "$2a$10$17L.XAhirq2flcIF2GnCResHiMMgLngltWWikkoaq2qsljvKDPaCe",
                  "role": "ADMIN",
                  "created_at": null
                }
              ]
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "813"
              },
              "etag": {
                "type": "string",
                "default": "W/\"32d-pPxjn3Ww3KlrK2aLCoy/sGzxKwU\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 20:36:35 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "No token, authorization denied!"
              }
            }
          }
        }
      }
    },
    "/user/delete/{id}": {
      "delete": {
        "description": "Delete a specific user",
        "summary": "http://localhost:3000/api/user/delete/:id",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/user/delete/:id",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      }
    },
    "/users/update_user/{id}": {
      "put": {
        "description": "Edit a specific user",
        "summary": "http://localhost:3000/api/users/update_user/:id",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/users/update_user/:id",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users~1update_user~1%3Aid"
            },
            "examples": {
              "application/json": {
                "users": [
                  {
                    "id": 4,
                    "name": "admin",
                    "email": "adwwwminbbjbjhbjhaaaa@mail.com",
                    "password": "$2a$10$G9aL5GMxEZnRc.Ng9.hxbuqojf3quSEDZpKeTzFKghVArOw6frwEe",
                    "role": "ADMIN",
                    "created_at": null
                  },
                  {
                    "id": 11,
                    "name": "admin",
                    "email": "adwwwminaaaa@mail.com",
                    "password": "$2a$10$kmlURv.RNIYegcJEdK0C6.MtCJ09FIDS.iGMrhwdNbFJkbKlVcmha",
                    "role": "ADMIN",
                    "created_at": null
                  },
                  {
                    "id": 10,
                    "name": "admin",
                    "email": "adminaaaa@mail.com",
                    "password": "$2a$10$I8wS/0x4/kKx24Y4/lNx3.IIfYt41R7Fo/tunI8BICHuPxPT2DwUG",
                    "role": "ADMIN",
                    "created_at": null
                  },
                  {
                    "id": 9,
                    "name": "Tony Vercetti",
                    "email": "tony@mail.com",
                    "password": "$2a$10$8hi7SETeQJwdSfPjHUIS5.L/pzE0OoGkvJE6XguagJ3ysQUR0IFnq",
                    "role": "ADMIN",
                    "created_at": null
                  },
                  {
                    "id": 8,
                    "name": "admin",
                    "email": "admin@mail.com",
                    "password": "$2a$10$17L.XAhirq2flcIF2GnCResHiMMgLngltWWikkoaq2qsljvKDPaCe",
                    "role": "ADMIN",
                    "created_at": null
                  }
                ],
                "successMsg": "User was updated!"
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "860"
              },
              "etag": {
                "type": "string",
                "default": "W/\"35c-pBcyS+RCC0az1qGNyUc+kFXlVHg\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 20:39:06 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          }
        }
      }
    }
  },
...
}